refactor: change name to camelCase
