+++ src/app/(root)/setting/page.tsx
@@ -1,9 +1,9 @@
 import { auth } from '@/lib/auth';
-import UpdateInformationForm from '@/components/setting-page/UpdateInformationForm';
+import UpdateInformationForm from '@/components/setting-page/updateInformationForm';
 import { redirect } from 'next/navigation';
-import Divider from '@/components/navigationBar/Divider';
-import VerificationStatus from '@/components/setting-page/VerificationStatus';
-import DeleteAccountButton from '@/components/setting-page/DeleteAccountButton';
+import Divider from '@/components/navigationBar/divider';
+import VerificationStatus from '@/components/setting-page/verificationStatus';
+import DeleteAccountButton from '@/components/setting-page/deleteAccountButton';
 import { AuthProvider } from '@/components/auth/authProvider';
 
 const page = async () => {
