+++ src/app/(sign)/register/page.tsx
@@ -1,4 +1,4 @@
-import { RegisterBox } from '@/components/register-page/RegisterBox';
+import { RegisterBox } from '@/components/register-page/registerBox';
 
 export default function Page() {
   return (
