+++ src/components/lesseeHome-page/RequestBtn.tsx
@@ -8,7 +8,7 @@ import {
   DialogTitle,
   DialogTrigger,
 } from '@/components/ui/dialog';
-import { RequestForm } from '@/components/lesseeHome-page/RequestForm';
+import { RequestForm } from '@/components/lesseeHome-page/requestForm';
 
 export function RequestBtn() {
   return (
