+++ src/components/lessorIncome-page/lessorDashboard.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import React from 'react';
-import { PropertyDetail } from '@/components/lesseeHome-page/PropertyDetail';
+import { PropertyDetail } from '@/components/lesseeHome-page/propertyDetail';
 import { mockData } from '@/mocks/mockProperty';
 import { PropertyI } from '@/types/property';
