+++ src/components/profileLesseeID-page/ProfileInfo.tsx
@@ -8,7 +8,7 @@ import {
 } from 'lucide-react';
 import { Card, CardContent, CardHeader } from '@/components/ui/card';
 
-import TextInCard from '@/components/profileLesseeID-page/TextInCard';
+import TextInCard from '@/components/profileLesseeID-page/textInCard';
 import { Badge } from '@/components/ui/badge';
 
 export default function ProfileInfo() {
