+++ src/components/register-page/RegisterForm.tsx
@@ -17,7 +17,7 @@ import {
   FormLabel,
   FormMessage,
 } from '@/components/ui/form';
-import InputWithIcon from '@/components/inputWithIcon/InputWithIcon';
+import InputWithIcon from '@/components/inputWithIcon/inputWithIcon';
 import { useRouter } from 'next/navigation';
 
 const formSchema = z
